feat: undelegate and close shutle and it's ata
